Added auth middleware
